Add wrapper around curl to make requests safely (#6)
The user agent that curl uses when making requests can be overridden
using using the "user-agent" setting in the `.curlrc` file, and setting
a custom user agent might break asdf-jq as it relies on a pretty-printed
JSON document to be returned when querying GitHub API for the list of
known releases.

GitHub API will only return a pretty-printed JSON document in the
response to a request made by curl (or any other client that uses the
same User-Agent header); otherwise, and for any other client, the JSON
document sent back will be minified (machine parsable).

Thus, ensure that requests asdf-jq makes always set the correct
User-Agent header, overriding any local or system-wide setting.

Signed-off-by: Krzysztof Wilczyński <kw@linux.com>